+++ zhaquirks/konke/button.py
@@ -11,12 +11,12 @@ from zigpy.zcl.clusters.general import (
     Scenes,
 )
 
-from .. import CustomCluster, CustomDevice, PowerConfigurationCluster
+from . import KonkeButtonRemote, KonkeOnOffCluster
+from .. import CustomDevice, PowerConfigurationCluster
 from ..const import (
     COMMAND,
     COMMAND_DOUBLE,
     COMMAND_HOLD,
-    COMMAND_ID,
     COMMAND_SINGLE,
     DEVICE_TYPE,
     DOUBLE_PRESS,
@@ -25,10 +25,8 @@ from ..const import (
     LONG_PRESS,
     MODELS_INFO,
     OUTPUT_CLUSTERS,
-    PRESS_TYPE,
     PROFILE_ID,
     SHORT_PRESS,
-    ZHA_SEND_EVENT,
 )
 
 KONKE_CLUSTER_ID = 0xFCC0
@@ -36,52 +34,8 @@ KONKE_CLUSTER_ID = 0xFCC0
 _LOGGER = logging.getLogger(__name__)
 
 
-class KonkeOnOffCluster(CustomCluster, OnOff):
-    """Konke OnOff cluster implementation."""
-
-    PRESS_TYPES = {0x0080: COMMAND_SINGLE, 0x0081: COMMAND_DOUBLE, 0x0082: COMMAND_HOLD}
-    cluster_id = 6
-    ep_attribute = "custom_on_off"
-    attributes = {}
-    server_commands = {}
-    client_commands = {}
-
-    def handle_cluster_general_request(self, header, args):
-        """Handle the cluster command."""
-        self.info(
-            "Konke general request - handle_cluster_general_request: header: %s - args: [%s]",
-            header,
-            args,
-        )
-
-        cmd = header.command_id
-        event_args = {
-            PRESS_TYPE: self.PRESS_TYPES.get(cmd, cmd),
-            COMMAND_ID: cmd,
-        }
-        self.listener_event(ZHA_SEND_EVENT, event_args[PRESS_TYPE], event_args)
-
-
-class KonkeButtonRemote(CustomDevice):
+class KonkeButtonRemote1(KonkeButtonRemote):
     """Konke 1-button remote device."""
-
-    def handle_message(self, profile, cluster, src_ep, dst_ep, message):
-        """Handle a device message."""
-        if (
-            profile == 260
-            and cluster == 6
-            and len(message) == 7
-            and message[0] == 0x08
-            and message[2] == 0x0A
-        ):
-            # use the 7th byte as command_id
-            new_message = bytearray(4)
-            new_message[0] = message[0]
-            new_message[1] = message[1]
-            new_message[2] = message[6]
-            new_message[3] = 0
-            message = type(message)(new_message)
-            super().handle_message(profile, cluster, src_ep, dst_ep, message)
 
     signature = {
         # <SimpleDescriptor endpoint=1 profile=260 device_type=2
@@ -133,24 +87,6 @@ class KonkeButtonRemote(CustomDevice):
 
 class KonkeButtonRemote2(CustomDevice):
     """Konke 1-button remote device 2nd variant."""
-
-    def handle_message(self, profile, cluster, src_ep, dst_ep, message):
-        """Handle a device message."""
-        if (
-            profile == 260
-            and cluster == 6
-            and len(message) == 7
-            and message[0] == 0x08
-            and message[2] == 0x0A
-        ):
-            # use the 7th byte as command_id
-            new_message = bytearray(4)
-            new_message[0] = message[0]
-            new_message[1] = message[1]
-            new_message[2] = message[6]
-            new_message[3] = 0
-            message = type(message)(new_message)
-            super().handle_message(profile, cluster, src_ep, dst_ep, message)
 
     signature = {
         # <SimpleDescriptor endpoint=1 profile=260 device_type=2
